+++ src/test/kotlin/com/bugsnag/android/gradle/ObjDumpLocationTest.kt
@@ -1,6 +1,6 @@
 package com.bugsnag.android.gradle
 
-import com.bugsnag.android.gradle.BugsnagGenerateNdkSoMappingTask.Companion.calculateObjDumpLocation
+import com.bugsnag.android.gradle.SharedObjectMappingFileFactory.calculateObjDumpLocation
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith